+++ source/client/pages.js
@@ -1,0 +1,9 @@
+import Catalog from '#custom/page/Catalog'
+import Component from '#custom/page/Component'
+import Document from '#custom/page/Document'
+import Home from '#custom/page/Home'
+import Isolate from '#custom/page/Isolate'
+import Section from '#custom/page/Section'
+
+
+export default { Catalog, Component, Document, Home, Isolate, Section }