+++ source/service/loader/assemblers/index.js
@@ -4,6 +4,7 @@ import operations from './operations.js'
 
 export default function (config, other)
 {
+    let { finalizeAsset } = config;
     let op = operations(config);
 
     let exec = proxet({}, key => async record => 
@@ -22,6 +23,13 @@ export default function (config, other)
         for (let routine of routines) if (!(await routine(record))) return null;
         /* eslint-enable no-await-in-loop */
         
+        if (record.tid && finalizeAsset)
+        {
+            let { tid, uid } = record;
+            record = finalizeAsset(record);
+            if (record) record = { ...record, tid, uid };
+        }
+
         return record;
     }